Updated for options pages & added microformats
Hey there, love ACF and love this add-on, nice work! 

Noticed an issue regarding the resources not getting added to options pages, so updated that

Finally, saw the issue about adding microformats to this add-on, and personally wanted that as well for my own project. So what I've done is made the default classes be the correct microformat classes based on the address spec, and then to take it a bit further, also account for the front-end address container wrap and whether people want to use a html5 `<address class="adr">` or straight html `<div class="adr">`. So I added an additional optional that allows people to choose which parent container to use.

Hope this helps! 